Refact runner sim to use inheritent component sim
